[refactor] FinaleProject 관련 로직 분류 - FinaleProject, FinaleTeam, FinaleProjectBoard 세 종류로 분류 - 분류 후 서버 작동 확인함 - 에러핸들링 방식 수정 필요
