+++ src/main/java/com/honeyboard/api/project/finale/mapper/FinaleProjectMapper.java
@@ -1,12 +1,8 @@
 package com.honeyboard.api.project.finale.mapper;
 
-import com.honeyboard.api.project.finale.model.request.FinaleProjectBoardRequest;
 import com.honeyboard.api.project.finale.model.request.FinaleProjectCreate;
-import com.honeyboard.api.project.finale.model.request.FinaleProjectTeamUpdate;
 import com.honeyboard.api.project.finale.model.request.FinaleProjectUpdate;
 import com.honeyboard.api.project.finale.model.response.*;
-import com.honeyboard.api.project.model.ProjectUserInfo;
-import com.honeyboard.api.project.model.TeamMemberInfo;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
@@ -16,12 +12,6 @@ interface FinaleProjectMapper {
 
     // 프로젝트 리스트 조회
     List<FinaleProjectList> selectFinaleProjectList();
-
-    // 팀이 없는 유저 조회
-    List<ProjectUserInfo> selectNoFinaleTeamUsers(int finaleProjectId);
-
-    // 팀 리스트 조회
-    List<FinaleTeamList> selectFinaleTeamList(int finaleProjectId);
 
     // 프로젝트 생성
     int insertFinaleProject(@Param("finaleProjectCreate")FinaleProjectCreate project);
@@ -38,42 +28,6 @@ interface FinaleProjectMapper {
 
     // 프로젝트 상세 조회
     FinaleProjectDetail selectFinaleProjectDetail(int finaleProjectId);
-
-    // 프로젝트 상세 팀 인원 조회
-    List<TeamMemberInfo> selectFinaleProjectDetailMembers(int finaleTeamId);
-
-    // 프로젝트 상세 내 게시글 조회
-    List<FinaleProjectBoardList> selectFinaleProjectDetailBoards(int finaleProjectId);
-
-    int updateFinaleProjectTeam(@Param("finaleProjectId") int finaleProjectId,
-                                @Param("finaleTeamId") int finaleTeamId,
-                                @Param("request") FinaleProjectTeamUpdate request);
-
-    FinaleProjectBoardDetail selectFinaleProjectBoardDetail(
-            @Param("finaleProjectId") int finaleProjectId,
-            @Param("boardId") int boardId
-    );
-
-    boolean checkFinaleProjectBoard(@Param("finaleProjectId") int finaleProjectId,
-                                    @Param("boardId") int boardId);
-
-    int selectFinaleTeamId(int finaleProjectId);
-
-    boolean checkTeamMember(@Param("teamId") int teamId,
-                            @Param("userId") int userId);
-
-    int insertFinaleProjectBoard(@Param("finaleProjectId") int finaleProjectId,
-                                 @Param("request") FinaleProjectBoardRequest request);
-
-    int updateFinaleProjectBoard(@Param("finaleProjectId") int finaleProjectId,
-                                 @Param("finaleProjectBoardId") int finaleProjectBoardId,
-                                 @Param("request") FinaleProjectBoardRequest request);
-
-    boolean checkBoardTeamMember(@Param("finaleProjectId") int finaleProjectId,
-                                 @Param("finaleProjectBoardId") int finaleProjectBoardId,
-                                 @Param("userId") int userId);
-
-    int selectLastInsertedBoardId();
 
     int selectLastInsertedId();
