+++ src/main/java/com/honeyboard/api/project/finale/service/FinaleProjectBoardService.java
@@ -1,9 +1,17 @@
 package com.honeyboard.api.project.finale.service;
 
 
-import java.util.List;
+import com.honeyboard.api.project.finale.model.request.FinaleProjectBoardRequest;
+import com.honeyboard.api.project.finale.model.response.FinaleProjectBoardDetail;
+import com.honeyboard.api.user.model.User;
+
 
 public interface FinaleProjectBoardService {
 
+    FinaleProjectBoardDetail getFinaleProjectBoardDetail(int finaleProjectId, int boardId);
+
+    int createFinaleProjectBoard(int finaleProjectId, FinaleProjectBoardRequest request, User currentUser);
+
+    boolean updateFinaleProjectBoard(int finaleProjectId, int finaleProjectBoardId, FinaleProjectBoardRequest request, User currentUser);
 
 }
